Move StopSignal to test_utils
