+++ chain/network/src/peer_manager.rs
@@ -1,16 +1,19 @@
 use std::cmp;
 use std::collections::{HashMap, HashSet};
+use std::net::SocketAddr;
+use std::pin::Pin;
 use std::sync::Arc;
 use std::time::Duration;
 
 use actix::actors::resolver::{ConnectAddr, Resolver};
 use actix::io::FramedWrite;
 use actix::{
-    Actor, ActorContext, ActorFuture, Addr, AsyncContext, Context, ContextFutureSpawner, Handler,
-    Recipient, Running, StreamHandler, SystemService, WrapFuture,
+    Actor, ActorFuture, Addr, AsyncContext, Context, ContextFutureSpawner, Handler, Recipient,
+    Running, StreamHandler, SystemService, WrapFuture,
 };
 use chrono::offset::TimeZone;
 use chrono::{DateTime, Utc};
+use futures::task::Poll;
 use futures::{future, Stream, StreamExt};
 use rand::{thread_rng, Rng};
 use tokio::net::{TcpListener, TcpStream};
@@ -33,15 +36,11 @@ use crate::types::{
     NetworkViewClientMessages, NetworkViewClientResponses, OutboundTcpConnect, PeerId,
     PeerIdOrHash, PeerList, PeerManagerRequest, PeerMessage, PeerRequest, PeerResponse, PeerType,
     PeersRequest, PeersResponse, Ping, Pong, QueryPeerStats, RawRoutedMessage, ReasonForBan,
-    RoutedMessage, RoutedMessageBody, RoutedMessageFrom, SendMessage, StopSignal, SyncData,
-    Unregister,
+    RoutedMessage, RoutedMessageBody, RoutedMessageFrom, SendMessage, SyncData, Unregister,
 };
 use crate::types::{
     NetworkClientMessages, NetworkConfig, NetworkRequests, NetworkResponses, PeerInfo,
 };
-use futures::task::Poll;
-use std::net::SocketAddr;
-use std::pin::Pin;
 
 /// How often to request peers from active peers.
 const REQUEST_PEERS_SECS: i64 = 60;
@@ -1376,20 +1375,6 @@ impl Handler<PeerRequest> for PeerManagerActor {
                 );
                 PeerResponse::NoResponse
             }
-        }
-    }
-}
-
-impl Handler<StopSignal> for PeerManagerActor {
-    type Result = ();
-
-    fn handle(&mut self, msg: StopSignal, ctx: &mut Self::Context) -> Self::Result {
-        debug!(target: "network", "Receive Stop Signal. Me: {:?}", self.peer_id);
-
-        if msg.should_panic {
-            panic!("Node crashed");
-        } else {
-            ctx.stop();
         }
     }
 }
