+++ chain/network/tests/stress_network.rs
@@ -8,8 +8,8 @@ use futures::FutureExt;
 use tracing::info;
 
 use near_client::{ClientActor, ViewClientActor};
-use near_network::test_utils::{convert_boot_nodes, open_port, GetInfo, WaitOrTimeout};
-use near_network::types::{NetworkViewClientMessages, NetworkViewClientResponses, StopSignal};
+use near_network::test_utils::{convert_boot_nodes, open_port, GetInfo, StopSignal, WaitOrTimeout};
+use near_network::types::{NetworkViewClientMessages, NetworkViewClientResponses};
 use near_network::{NetworkClientResponses, NetworkConfig, PeerManagerActor};
 use near_primitives::block::WeightAndScore;
 use near_primitives::test_utils::init_test_logger_allow_panic;
